ClimaCache

Plant

ClimaCache.HyperspectralAbsorption — Type

mutable struct HyperspectralAbsorption{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflection and transmittance.

Fields

K_ANT::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of anthocynanin [-]

K_BROWN::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of senescent material (brown pigments) [-]

K_CAB::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of chlorophyll a and b [-]

K_CAR_V::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of violaxanthin carotenoid [-]

K_CAR_Z::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of zeaxanthin carotenoid [-]

K_CBC::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of carbon-based constituents [-]

K_H₂O::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of water [-]

K_LMA::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of dry matter [-]

K_PRO::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of protein [-]

K_PS::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of PS I and II [-]

NR::Vector{FT} where FT<:AbstractFloat

Refractive index [-]

ClimaCache.HyperspectralAbsorption — Method

HyperspectralAbsorption{FT}(wls::WaveLengthSet; opti::String=OPTI_2021) where {FT<:AbstractFloat}

Constructor for HyperspectralAbsorption, given

wls WaveLengthSet type structure

opti Path to leaf optical properties

Examples

ha = HyperspectralAbsorption{FT}();
ha = HyperspectralAbsorption{FT}(WaveLengthSet{FT}(collect(400:50:2400)));
ha = HyperspectralAbsorption{FT}(WaveLengthSet{FT}(collect(400:50:2400)); opti=ClimaCache.OPTI_2017);

ClimaCache.LeafBiophysics — Type

mutable struct LeafBiophysics{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflection and transmittance.

Fields

MESOPHYLL_N::AbstractFloat

Leaf mesophyll structural parameter that describes mesophyll reflection and transmittance

NDUB::Int64

Doubling adding layers

ant::AbstractFloat

Anthocynanin content [ug cm⁻²]

brown::AbstractFloat

Senescent material (brown pigments) fraction [-]

cab::AbstractFloat

Chlorophyll a and b content [ug cm⁻²]

car::AbstractFloat

Carotenoid content [ug cm⁻²]

cbc::AbstractFloat

Carbon-based constituents in lma [g cm⁻²]

f_zeax::Any

Zeaxanthin fraction in Carotenoid (1=all Zeaxanthin, 0=all Violaxanthin) [-]

l_H₂O::AbstractFloat

Equivalent water thickness [cm]

lma::AbstractFloat

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

pro::AbstractFloat

Protein content in lma (pro = lma - cbc) [g cm⁻²]

k_all::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of all materials

mat_b::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix backwards [-]

mat_f::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix forwards [-]

α_cab::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll [-]

α_cabcar::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll+Carotenoid [-]

α_SW::Vector{FT} where FT<:AbstractFloat

Shortwave absorption, 1 .- ρSW .- τSW [-]

ρ_LW::AbstractFloat

Broadband thermal reflectance, related to blackbody emittance [-]

ρ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf reflectance [-]

τ_LW::AbstractFloat

Broadband thermal transmission, related to blackbody emittance [-]

τ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf transmission [-]

ClimaCache.LeafBiophysics — Method

LeafBiophysics{FT}(wls::WaveLengthSet) where {FT<:AbstractFloat}

Constructor for LeafBiophysics, given

wls WaveLengthSet type structure

Examples

lbio = LeafBiophysics{FT}();
lbio = LeafBiophysics{FT}(WaveLengthSet{FT}(collect(400:50:2400)));

ClimaCache.AbstractPhotosynthesisSystem — Type

abstract type AbstractPhotosynthesisSystem{FT<:AbstractFloat}

Hierachy of AbstractSoilVC:

C₃VJPSystem

ClimaCache.C₃VJPSystem — Type

C₃VJPSystem{FT}(; v_max25::Number = 50, j_max25::Number = 83.5, r_d25::Number = 0.75) where {FT<:AbstractFloat}

Constructor for C₃VJPSystem, given

v_max25 Maximal carboxylation rate at 298.15 K

j_max25 Maximal electron transport rate at 298.15 K

r_d25 Respiration rate at 298.15 K

Examples

c3 = C₃VJPSystem{Float64}();
c3 = C₃VJPSystem{Float64}(v_max25 = 30, j_max25 = 50, r_d25 = 1);

ClimaCache.C₃VJPSystem — Method

C₃VJPSystem{FT}(; v_max25::Number = 50, j_max25::Number = 83.5, r_d25::Number = 0.75) where {FT<:AbstractFloat}

Constructor for C₃VJPSystem, given

v_max25 Maximal carboxylation rate at 298.15 K

j_max25 Maximal electron transport rate at 298.15 K

r_d25 Respiration rate at 298.15 K

Examples

c3 = C₃VJPSystem{Float64}();
c3 = C₃VJPSystem{Float64}(v_max25 = 30, j_max25 = 50, r_d25 = 1);

Radiation

ClimaCache.HyperspectralRadiation — Type

mutable struct HyperspectralRadiation{FT<:AbstractFloat}

Structure that stores hyperspectral radiation information

Fields

e_direct::Vector{FT} where FT<:AbstractFloat

Direct radiation [mW m⁻² nm⁻¹]

e_diffuse::Vector{FT} where FT<:AbstractFloat

Diffuse radiation [mW m⁻² nm⁻¹]

ClimaCache.HyperspectralRadiation — Method

HyperspectralRadiation{FT}(wls::WaveLengthSet = WaveLengthSet{FT}(), file::String = FILE_SUN)

Constructor for HyperspectralRadiation, given

wls WaveLengthSet type struct that defines wavelength settings

file File path to solar radiation setting, default is ClimaCache.FILE_SUN

Examples

rad = HyperspectralRadiation{FT}();
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)));
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)); file = "");
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)); file = ClimaCache.FILE_SUN);

ClimaCache.WaveLengthSet — Type

mutable struct WaveLengthSet{FT<:AbstractFloat}

Structure that stores wave length information.

Fields

IΛ_NIR::Vector{Int64}

Indicies of Λ_NIR in Λ

IΛ_PAR::Vector{Int64}

Indicies of Λ_PAR in Λ

IΛ_SIF::Vector{Int64}

Indicies of Λ_SIF in Λ

IΛ_SIFE::Vector{Int64}

Indicies of Λ_SIFE in Λ

NΛ::Int64

Number of wavelength bins

NΛ_PAR::Int64

Number of wavelength bins for PAR

NΛ_SIF::Int64

Number of wavelength bins for SIF

NΛ_SIFE::Int64

Number of wavelength bins for SIF excitation

WL_NIR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for NIR [nm]

WL_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for PAR [nm]

WL_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF emission [nm]

WL_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF excitation [nm]

SΛ::Vector{FT} where FT<:AbstractFloat

Standard wavelength (boundaries) [nm]

ΔΛ::Vector{FT} where FT<:AbstractFloat

Differential wavelength [nm]

ΔΛ_PAR::Vector{FT} where FT<:AbstractFloat

Differential wavelength for PAR [nm]

ΔΛ_SIFE::Vector{FT} where FT<:AbstractFloat

Differential wavelength for SIF excitation [nm]

Λ::Vector{FT} where FT<:AbstractFloat

Wavelength (bins) [nm]

Λ_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength bins for PAR [nm]

Λ_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF [nm]

Λ_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF excitation [nm]

ClimaCache.WaveLengthSet — Method

WaveLengthSet{FT}(swl::Vector=WAVELENGTHS; opti::String=OPTI_2021) where {FT<:AbstractFloat}

Constructor for WaveLengthSet, given

swl Standard wave length boundaries [nm], default is ClimaCache.WAVELENGTHS

opti Optical properties file path, default is ClimaCache.OPTI_2021

Examples

wls = WaveLengthSet{FT}();
wls = WaveLengthSet{FT}(collect(400:5:2500));
wls = WaveLengthSet{FT}(collect(400:5:2500); opti=ClimaCache.OPTI_2017);

Soil

ClimaCache.AbstractSoilVC — Type

abstract type AbstractSoilVC{FT<:AbstractFloat}

Hierachy of AbstractSoilVC:

BrooksCorey

VanGenuchten

ClimaCache.BrooksCorey — Type

struct BrooksCorey{FT<:AbstractFloat} <: ClimaCache.AbstractSoilVC{FT<:AbstractFloat}

Brooks Corey soil parameters

Fields

TYPE::String

Soil type

B::AbstractFloat

Soil b

Ψ_SAT::AbstractFloat

Potential at saturation [MPa]

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

ClimaCache.VanGenuchten — Type

struct VanGenuchten{FT<:AbstractFloat} <: ClimaCache.AbstractSoilVC{FT<:AbstractFloat}

van Genuchten soil parameters

Fields

TYPE::String

Soil type

Α::AbstractFloat

Soil α is related to the inverse of the air entry suction, α > 0

N::AbstractFloat

Soil n is Measure of the pore-size distribution

M::AbstractFloat

Soil m = 1 - 1/n

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

ClimaCache.VanGenuchten — Method

VanGenuchten{FT}(name::String, α::Number, n::Number, θ_sat::Number, θ_res::Number) where {FT<:AbstractFloat}

Constructor for VanGenuchten, given

name Soil type name

α van Genuchten α

n van Genuchten n

θ_sat Saturated soil volumetric water content

θ_res Residual soil volumetric water content

Examples

vg = VanGenuchten{FT}("Test", FT(100), FT(2), FT(0.5), FT(0.1));

ClimaCache.VanGenuchten — Method

VanGenuchten{FT}(name::String) where {FT<:AbstractFloat}

Constructor for VanGenuchten, given

name Soil type name, need to be Sand, Loamy Sand, Sandy Loam, Loam, Sandy Clay Loam, Silt Loam, Silt (default), Clay Loam, Silty Clay Loam, Sandy Clay, Silty Clay, and Clay.

Examples

vg = VanGenuchten{FT}("");
vg = VanGenuchten{FT}("Loam");