ClimaCache

Air

ClimaCache.AirLayer — Type

mutable struct AirLayer{FT<:AbstractFloat}

Structure that stores air layer information

Fields

P_AIR::AbstractFloat

Atmospheric pressure [Pa]

P_O₂::AbstractFloat

O₂ partial pressure [Pa]

p_CO₂::AbstractFloat

CO₂ partial pressure [Pa]

p_H₂O::AbstractFloat

H₂O partial pressure [Pa]

t::AbstractFloat

Temperature

wind::AbstractFloat

Wind speed [m s⁻¹]

p_H₂O_sat::AbstractFloat

Saturated H₂O partial pressure [Pa]

rh::AbstractFloat

relative humidity

ClimaCache.AirLayer — Method

AirLayer{FT}() where {FT<:AbstractFloat}

Constructor for AirLayer

Examples

air = AirLayer{Float64}();

Plant

Hydraulics

ClimaCache.AbstractPVCurve — Type

abstract type AbstractPVCurve{FT<:AbstractFloat}

Hierarchy of AbstractPVCurve:

LinearPVCurve

SegmentedPVCurve

ClimaCache.LinearPVCurve — Type

mutable struct LinearPVCurve{FT<:AbstractFloat} <: ClimaCache.AbstractPVCurve{FT<:AbstractFloat}

Struct that contains information for linear PV curve

Fields

K_REFILL::AbstractFloat

Conductance for refilling (relative to maximum) [MPa⁻¹ s⁻¹]

SLOPE::AbstractFloat

Slope of the linear PV curve (relative to maximum) [MPa⁻¹]

ClimaCache.LinearPVCurve — Method

LinearPVCurve{FT}() where {FT<:AbstractFloat}

Constructor for LinearPVCurve

Examples

pvc = LinearPVCurve{Float64}();

ClimaCache.SegmentedPVCurve — Type

mutable struct SegmentedPVCurve{FT} <: ClimaCache.AbstractPVCurve{FT}

Struct that contains information for segmented PV curve

Fields

C_ALL::Any

n_o / maximum V [mol m⁻³]

K_REFILL::Any

Conductance for refilling (relative to maximum) [MPa⁻¹ s⁻¹]

RWC_APO::Any

Apoplastic water content relative to maximum water volume

RWC_TLP::Any

Relative water content at turgor loss point

Ε_BULK::Any

Bulk modulus of elasticity [MPa]

ClimaCache.SegmentedPVCurve — Method

SegmentedPVCurve{FT}() where {FT<:AbstractFloat}

Constructor for SegmentedPVCurve

Examples

pvc = SegmentedPVCurve{Float64}();

ClimaCache.AbstractHydraulicSystem — Type

abstract type AbstractHydraulicSystem{FT<:AbstractFloat}

Hierarchy of AbstractHydraulicSystem:

LeafHydraulics

RootHydraulics

ClimaCache.AbstractXylemVC — Type

abstract type AbstractXylemVC{FT<:AbstractFloat}

Hierarchy of AbstractXylemVC:

LogisticVC

PowerVC

WeibullVC

ComplexVC

ClimaCache.LogisticVC — Type

mutable struct LogisticVC{FT<:AbstractFloat} <: ClimaCache.AbstractXylemVC{FT<:AbstractFloat}

Modified logistic function for vulnerability curve

Fields

A::AbstractFloat

Multiplier to exponential component

B::AbstractFloat

Multiplier to pressure [MPa⁻¹]

ClimaCache.PowerVC — Type

mutable struct PowerVC{FT<:AbstractFloat} <: ClimaCache.AbstractXylemVC{FT<:AbstractFloat}

Power function for vulnerability curve

Fields

A::AbstractFloat

Multiplier to power component [MPa⁻ᵇ]

B::AbstractFloat

Power to pressure

ClimaCache.WeibullVC — Type

mutable struct WeibullVC{FT<:AbstractFloat} <: ClimaCache.AbstractXylemVC{FT<:AbstractFloat}

Weibull cumulative distribution function for vulnerability curve

Fields

B::AbstractFloat

Numerator in the exponential component [MPa]

C::AbstractFloat

Power to pressure component

ClimaCache.ComplexVC — Type

mutable struct ComplexVC{FT} <: ClimaCache.AbstractXylemVC{FT}

A complex struct for segmented vulnerability curve such as dual Weibull function

Fields

PS::Vector

Percentages of each VC component

VCS::Union{Array{ClimaCache.AbstractXylemVC{FT}, 1}, Array{ClimaCache.LogisticVC{FT}, 1}, Array{ClimaCache.PowerVC{FT}, 1}, Array{ClimaCache.WeibullVC{FT}, 1}} where FT

Vector of vulnerability curve components

ClimaCache.AbstractFlowProfile — Type

abstract type AbstractFlowProfile{FT<:AbstractFloat}

Hierarchy of AbstractFlowProfile:

NonSteadyStateFlow

SteadyStateFlow

ClimaCache.NonSteadyStateFlow — Type

mutable struct NonSteadyStateFlow{FT} <: ClimaCache.AbstractFlowProfile{FT}

Struct that contains stem hydraulic system flow rates at non-steady state

Fields

f_buffer::Vector

Vector of buffer water flow [mol m⁻²]

f_element::Vector

Vector of xylem water flow [mol m⁻²]

f_in::Any

Flow rate in [mol s⁻¹] or [mol m⁻² s⁻¹] (for leaf)

f_out::Any

Flow rate out [mol s⁻¹] or [mol m⁻² s⁻¹] (for leaf)

f_sum::Vector

Vector of sum buffer water flow [mol m⁻²]

ClimaCache.NonSteadyStateFlow — Method

NonSteadyStateFlow{FT}(N::Int, isleaf::Bool = true) where {FT<:AbstractFloat}

Construct a non-steady state flow struct, given

N Number of buffer rates from capaciatance (always 1 for Leaf)

isleaf Bool to indicate if the organ is a leaf

ClimaCache.SteadyStateFlow — Type

mutable struct SteadyStateFlow{FT} <: ClimaCache.AbstractFlowProfile{FT}

Struct that contains stem hydraulic system flow rates at steady state

Fields

flow::Any

Flow rate [mol s⁻¹] or [mol m⁻² s⁻¹] (for leaf)

ClimaCache.LeafHydraulics — Type

mutable struct LeafHydraulics{FT} <: ClimaCache.AbstractHydraulicSystem{FT}

Struct that contains leaf hydraulic system

Fields

AREA::Any

Leaf area

FLOW::ClimaCache.AbstractFlowProfile{FT} where FT

Flow profile

K_OX::Any

Maximal extra-xylary hydraulic conductance [mol s⁻¹ MPa⁻¹ m⁻²]

K_SLA::Any

Maximal leaf xylem hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

N::Int64

Number of xylem slices

PVC::ClimaCache.AbstractPVCurve{FT} where FT

Pressure volume curve for storage

V_MAXIMUM::Any

Total capaciatance at Ψ = 0 [mol m⁻²]

VC::ClimaCache.AbstractXylemVC{FT} where FT

Vulnerability curve

p_dos::Any

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

p_leaf::Any

Leaf end water pressure at air-water interface [MPa]

p_storage::Any

Pressure of storage

p_ups::Any

Leaf xylem water pressure at the leaf base (upstream) [MPa]

v_storage::Any

Current capaciatance at Ψ_leaf [mol m⁻²]

k_history::Vector

Vector of leaf kr history per element

p_element::Vector

Vector of xylem water pressure [MPa]

p_history::Vector

Vector of xylem water pressure history (normalized to 298.15 K) [MPa]

ClimaCache.LeafHydraulics — Method

LeafHydraulics{FT}(N::Int = 5; area::Number = 1500, k_ox::Number = 100, k_sla::Number = 0.04, v_max::Number = 20, ssm::Bool = true) where {FT<:AbstractFloat}

Constructor for leaf hydraulic system, given

N Number of xylem slices in the system, default is 5

area Leaf area

k_ox Maximum extraxylary hydraulic conductance per leaf area

k_sla Maximum leaf xylem hydraulic conductance per leaf area

v_max Total water capacitance at Ψ = 0 per leaf area

ssm Whether the flow rate is at steady state

Examples

lhs = LeafHydraulics{Float64}();
lhs = LeafHydraulics{Float64}(N = 5);
lhs = LeafHydraulics{Float64}(N = 5; area = 20, k_ox = 50, k_sla = 0.1, v_max = 20, ssm = true);

ClimaCache.RootHydraulics — Type

mutable struct RootHydraulics{FT} <: ClimaCache.AbstractHydraulicSystem{FT}

Struct that contains root hydraulic system

Fields

AREA::Any

Root cross-section area [m²]

FLOW::ClimaCache.AbstractFlowProfile{FT} where FT

Flow profile

K_MAX::Any

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

K_RHIZ::Any

Rhizosphere conductance [mol s⁻¹ MPa⁻¹]

K_X::Any

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

L::Any

Length [m]

N::Int64

Number of xylem slices

PVC::ClimaCache.AbstractPVCurve{FT} where FT

Pressure volume curve for storage

SH::ClimaCache.AbstractSoilVC{FT} where FT

Soil hydraulics

V_MAXIMUM::Vector

Maximal storage per element [mol]

VC::ClimaCache.AbstractXylemVC{FT} where FT

Vulnerability curve

ΔH::Any

Root z difference [m]

p_dos::Any

Xylem water pressure at the downstream end of xylem [MPa]

p_rhiz::Any

Xylem-rhizosphere interface water pressure [MPa]

p_storage::Vector

Pressure of storage per element

p_ups::Any

Soil matrix potential [MPa]

v_storage::Vector

Storage per element [mol]

ψ_osm::Any

Soil osmotic potential at 298.15 K `[MPa]

k_history::Vector

Vector of leaf kr history per element

p_element::Vector

Vector of xylem water pressure [MPa]

p_history::Vector

Vector of xylem water pressure history (normalized to 298.15 K) [MPa]

ClimaCache.RootHydraulics — Method

RootHydraulics{FT}(N::Int = 5; area::Number = 1, k_x::Number = 25, Δh::Number = 1, Δl::Number = 1, ssm::Bool = true) where {FT<:AbstractFloat}

Constructor for root hydraulic system, given

N Number of xylem slices in the system, default is 5

area Root crosssection area

k_x Maximum root xylem hydraulic conductivity per crosssection area per root length

Δh Root depth

Δl Root length

ssm Whether the flow rate is at steady state

Examples

rhs = RootHydraulics{Float64}();
rhs = RootHydraulics{Float64}(N = 5);
rhs = RootHydraulics{Float64}(N = 5; area = 1, k_x = 50, Δh = 1, Δl = 2, ssm = true);

ClimaCache.StemHydraulics — Type

mutable struct StemHydraulics{FT} <: ClimaCache.AbstractHydraulicSystem{FT}

Struct that contains stem hydraulic system

Fields

AREA::Any

Root cross-section area [m²]

FLOW::ClimaCache.AbstractFlowProfile{FT} where FT

Flow profile

K_MAX::Any

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

K_X::Any

Maximal xylem hydraulic conductivity (per root depth) [mol s⁻¹ MPa⁻¹ m⁻²]

L::Any

Length [m]

N::Int64

Number of xylem slices

PVC::ClimaCache.AbstractPVCurve{FT} where FT

Pressure volume curve for storage

V_MAXIMUM::Vector

Maximal storage per element [mol]

VC::ClimaCache.AbstractXylemVC{FT} where FT

Vulnerability curve

ΔH::Any

Root z difference [m]

p_dos::Any

Xylem water pressure at the downstream end of xylem [MPa]

p_storage::Vector

Pressure of storage per element

p_ups::Any

Soil matrix potential [MPa]

v_storage::Vector

Storage per element [mol]

k_history::Vector

Vector of leaf kr history per element

p_element::Vector

Vector of xylem water pressure [MPa]

p_history::Vector

Vector of xylem water pressure history (normalized to 298.15 K) [MPa]

ClimaCache.StemHydraulics — Method

StemHydraulics{FT}(N::Int = 5; area::Number = 1, k_x::Number = 25, Δh::Number = 1, Δl::Number = 1, ssm::Bool = true) where {FT<:AbstractFloat}

Constructor for stem hydraulic system, given

N Number of xylem slices in the system, default is 5

area Root crosssection area

k_x Maximum stem xylem hydraulic conductivity per crosssection area per stem length

Δh Stem height

Δl Stem length

ssm Whether the flow rate is at steady state

Examples

rhs = StemHydraulics{Float64}();
rhs = StemHydraulics{Float64}(N = 5);
rhs = StemHydraulics{Float64}(N = 5; area = 1, k_x = 50, Δh = 2, ssm = true);

Leaf Level

ClimaCache.Leaf — Type

mutable struct Leaf{FT<:AbstractFloat}

Structure to save leaf parameters

Fields

BIO::ClimaCache.LeafBiophysics

LeafBiophysics type leaf biophysical parameters

HS::ClimaCache.LeafHydraulics{FT} where FT<:AbstractFloat

LeafHydraulics type leaf hydraulic system

PRC::ClimaCache.AbstractReactionCenter

AbstractReactionCenter type photosynthesis reaction center

PSM::ClimaCache.AbstractPhotosynthesisModel

AbstractPhotosynthesisModel type photosynthesis model

apar::AbstractFloat

Absorbed photosynthetically active radiation [μmol m⁻² s⁻¹]

g_H₂O_s::AbstractFloat

Stomatal conductance to water vapor [mol m⁻² s⁻¹]

t::AbstractFloat

Current leaf temperature

g_CO₂::AbstractFloat

Total leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

g_CO₂_b::AbstractFloat

Boundary leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

p_CO₂_i::AbstractFloat

Leaf internal CO₂ partial pressure [Pa]

p_CO₂_s::AbstractFloat

Leaf surface CO₂ partial pressure [Pa]

p_H₂O_sat::AbstractFloat

Saturation H₂O vapor pressure, need to update with temperature and leaf water pressure [Pa]

_t::AbstractFloat

Last leaf temperature. If different from t, then make temperature correction

ClimaCache.Leaf — Method

Leaf{FT}(psm::String, wls::WaveLengthSet{FT} = WaveLengthSet{FT}(); colimit::Bool = false, ssm::Bool = true) where {FT<:AbstractFloat}

Constructor for Leaf, given

psm Photosynthesis model type, must be C3, C3Cytochrome, or C4

wls WaveLengthSet type structure that determines the dimensions of leaf parameters

ssm Whether the flow rate is at steady state

Examples

leaf_c3 = Leaf{Float64}("C3");
leaf_c4 = Leaf{Float64}("C4");
leaf_cy = Leaf{Float64}("C3Cytochrome");
leaf_c3 = Leaf{Float64}("C3"; colimit = true);
leaf_c4 = Leaf{Float64}("C4"; colimit = true);
leaf_cy = Leaf{Float64}("C3Cytochrome"; colimit = true);
wls = WaveLengthSet{FT}(collect(400:10:2500));
leaf_c3 = Leaf{Float64}("C3", wls);
leaf_c4 = Leaf{Float64}("C4", wls);
leaf_cy = Leaf{Float64}("C3Cytochrome", wls);

ClimaCache.LeafBiophysics — Type

mutable struct LeafBiophysics{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflectance and transmittance.

Fields

MESOPHYLL_N::AbstractFloat

Leaf mesophyll structural parameter that describes mesophyll reflectance and transmittance

NDUB::Int64

Doubling adding layers

ant::AbstractFloat

Anthocyanin content [μg cm⁻²]

brown::AbstractFloat

Senescent material (brown pigments) fraction [-]

cab::AbstractFloat

Chlorophyll a and b content [μg cm⁻²]

car::AbstractFloat

Carotenoid content [μg cm⁻²]

cbc::AbstractFloat

Carbon-based constituents in lma [g cm⁻²]

f_zeax::AbstractFloat

Zeaxanthin fraction in Carotenoid (1=all Zeaxanthin, 0=all Violaxanthin) [-]

l_H₂O::AbstractFloat

Equivalent water thickness [cm]

lma::AbstractFloat

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

pro::AbstractFloat

Protein content in lma (pro = lma - cbc) [g cm⁻²]

k_all::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of all materials

mat_b::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix backwards [-]

mat_f::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix forwards [-]

α_cab::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll [-]

α_cabcar::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll+Carotenoid [-]

α_SW::Vector{FT} where FT<:AbstractFloat

Shortwave absorption, 1 .- ρSW .- τSW [-]

ρ_LW::AbstractFloat

Broadband thermal reflectance, related to blackbody emittance [-]

ρ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf reflectance [-]

τ_LW::AbstractFloat

Broadband thermal transmission, related to blackbody emittance [-]

τ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf transmission [-]

ClimaCache.LeafBiophysics — Method

LeafBiophysics{FT}(wls::WaveLengthSet{FT} = WaveLengthSet{FT}()) where {FT<:AbstractFloat}

Constructor for LeafBiophysics, given

wls WaveLengthSet type structure

Examples

lbio = LeafBiophysics{Float64}();
lbio = LeafBiophysics{Float64}(WaveLengthSet{Float64}(collect(400:50:2400)));

ClimaCache.VanDerTolFluorescenceModel — Type

struct VanDerTolFluorescenceModel{FT<:AbstractFloat}

Structure that stores van der Tol et al. (2014) fluorescence model parameters.

Fields

K_0::AbstractFloat

Fitting parameter K_0

K_A::AbstractFloat

Fitting parameter α

K_B::AbstractFloat

Fitting parameter β

ClimaCache.VanDerTolFluorescenceModel — Method

VanDerTolFluorescenceModel{FT}(drought::Bool = true) where {FT<:AbstractFloat}

Constructor for VanDerTolFluorescenceModel fluorescence model, given

drought If true, use parameters trained from drought stressed plant. Default is true.

Examples

vdt = VanDerTolFluorescenceModel{Float64}();
vdt = VanDerTolFluorescenceModel{Float64}(false);

ClimaCache.AbstractReactionCenter — Type

abstract type AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Hierarchy of the AbstractReactionCenter

VJPReactionCenter

CytochromeReactionCenter

ClimaCache.VJPReactionCenter — Type

mutable struct VJPReactionCenter{FT<:AbstractFloat} <: ClimaCache.AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Fields

F_PSII::AbstractFloat

Fraction of absorbed light used by PSII ETR

FLM::ClimaCache.VanDerTolFluorescenceModel

Fluorescence model

K_D::AbstractFloat

Rate constant for thermal dissipation

K_F::AbstractFloat

Rate constant for fluorescence

K_P_MAX::AbstractFloat

Maximal rate constant for photochemistry

Φ_PSII_MAX::AbstractFloat

max PSII yield (knpqr=0, all RC open)

k_npq_rev::AbstractFloat

Reversible NPQ rate constant (initially zero)

k_npq_sus::AbstractFloat

Sustained NPQ rate constant (for seasonal changes, default is zero)

k_p::AbstractFloat

Rate constant for photochemistry

npq::AbstractFloat

Non-Photochemical quenching

ϕ_f::AbstractFloat

Fluorescence yield

ϕ_p::AbstractFloat

Photochemical yield

f_m::AbstractFloat

Dark adapted yield (Kp=0)

f_m′::AbstractFloat

Light adapted yield (Kp=0)

f_o::AbstractFloat

Dark-adapted fluorescence yield (Kp=max)

f_o′::AbstractFloat

Light-adapted fluorescence yield in the dark (Kp=max)

q_e::AbstractFloat

Energy quenching

q_p::AbstractFloat

Photochemical quenching

ClimaCache.VJPReactionCenter — Method

VJPReactionCenter{FT}() where {FT<:AbstractFloat}

Constructor of VJPReactionCenter

Examples

rc = VJPReactionCenter{Float64}();

ClimaCache.CytochromeReactionCenter — Type

mutable struct CytochromeReactionCenter{FT<:AbstractFloat} <: ClimaCache.AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Fields

F_PSI::AbstractFloat

Fraction of absorbed light used by PSI ETR

K_D::AbstractFloat

Rate constant of consititutive heat loss from the antennae [ns⁻¹]

K_F::AbstractFloat

Rate constant of fluorescence [ns⁻¹]

K_PSI::AbstractFloat

Rate constant of photochemistry for PS I [ns⁻¹]

K_PSII::AbstractFloat

Rate constant of photochemistry for PS II [ns⁻¹]

K_U::AbstractFloat

Rate constant of excitation sharing for PS II [ns⁻¹]

K_X::AbstractFloat

Rate constant of regulated heat loss via oxidized PS I center [s⁻¹]

Φ_PSI_MAX::AbstractFloat

Maximal PS I photochemical yield

ϵ_1::AbstractFloat

Weight factor that PSI fluorescence reaches sensor (after reabsorption)

ϵ_2::AbstractFloat

Weight factor that PSII fluorescence reaches sensor (after reabsorption)

ϕ_f::AbstractFloat

Fluorescence yield

ϕ_p::AbstractFloat

Photochemical yield

ClimaCache.CytochromeReactionCenter — Method

CytochromeReactionCenter{FT}() where {FT<:AbstractFloat}

Constructor of CytochromeReactionCenter

Examples

rc = CytochromeReactionCenter{Float64}();

ClimaCache.AbstractPhotosynthesisModel — Type

abstract type AbstractPhotosynthesisModel{FT<:AbstractFloat}

Hierarchy of AbstractPhotosynthesisModel:

C3CytochromeModel

C3VJPModel

C4VJPModel

ClimaCache.C3CytochromeModel — Type

mutable struct C3CytochromeModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C3 Cytochrome photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

COLIMIT_J::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for J

EFF_1::AbstractFloat

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

EFF_2::AbstractFloat

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

TD_KC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kc temperature dependency

TD_KO::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Ko temperature dependency

TD_KQ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kq temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_Γ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Γ* temperature dependency

TD_ΗC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Η_C temperature dependency

TD_ΗL::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Η_L temperature dependency

b₆f::AbstractFloat

Total concentration of Cytochrome b₆f [μmol m⁻²]

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25_ww::AbstractFloat

Well watered maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol e⁻ m⁻² s⁻¹]

j_psi::AbstractFloat

PSI electron transport rate after colimitation

k_c::AbstractFloat

RubisCO coefficient Kc [Pa]

k_m::AbstractFloat

Michaelis-Menten's coefficient [Pa]

k_o::AbstractFloat

RubisCO coefficient Ko [Pa]

k_q::AbstractFloat

Maximal turnover rate of Cytochrome b₆f [e⁻ s⁻¹]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

v_qmax::AbstractFloat

Maximal Cytochrome b₆f activity [μmol e⁻ m⁻² s⁻¹]

η::AbstractFloat

ratio between JP700 and JP680

η_c::AbstractFloat

Coupling efficiency of cyclic electron flow [mol ATP mol⁻¹ e⁻]

η_l::AbstractFloat

Coupling efficiency of linear electron flow [mol ATP mol⁻¹ e⁻]

γ_star::AbstractFloat

CO₂ compensation point with the absence of Rd [Pa]

ClimaCache.C3CytochromeModel — Method

C3CytochromeModel{FT}(; v_cmax25::Number = 50, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C3CytochromeModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for ac, aj, and a_p

Examples

cy = C3CytochromeModel{Float64}();
cy = C3CytochromeModel{Float64}(v_cmax25 = 30, r_d25 = 1, colimit = true);

ClimaCache.C3VJPModel — Type

mutable struct C3VJPModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C3 photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

COLIMIT_J::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for J

EFF_1::AbstractFloat

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

EFF_2::AbstractFloat

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

TD_JMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Jmax temperature dependency

TD_KC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kc temperature dependency

TD_KO::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Ko temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_Γ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Γ* temperature dependency

j_max25::AbstractFloat

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25_ww::AbstractFloat

Well watered maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j::AbstractFloat

Electron transport [μmol m⁻² s⁻¹]

j_max::AbstractFloat

Maximal electron transport rate at leaf temperature [μmol m⁻² s⁻¹]

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

k_c::AbstractFloat

RubisCO coefficient Kc [Pa]

k_m::AbstractFloat

Michaelis-Menten's coefficient [Pa]

k_o::AbstractFloat

RubisCO coefficient Ko [Pa]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

γ_star::AbstractFloat

CO₂ compensation point with the absence of Rd [Pa]

ClimaCache.C3VJPModel — Method

C3VJPModel{FT}(; v_cmax25::Number = 50, j_max25::Number = 83.5, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C3VJPModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

j_max25 Maximal electron transport rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for j and ac, aj, and a_p

Examples

c3 = C3VJPModel{Float64}();
c3 = C3VJPModel{Float64}(v_cmax25 = 30, j_max25 = 50, r_d25 = 1, colimit = true);

ClimaCache.C4VJPModel — Type

mutable struct C4VJPModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C4 photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

TD_KPEP::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kpep temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_VPMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vpmax temperature dependency

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25_ww::AbstractFloat

Well watered maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

v_pmax25::AbstractFloat

Maximal PEP carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

k_pep::AbstractFloat

PEP coefficient Kpep [Pa]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

v_pmax::AbstractFloat

Maximal PEP carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

ClimaCache.C4VJPModel — Method

C4VJPModel{FT}(; v_cmax25::Number = 50, v_pmax25::Number = 50, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C4VJPModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

v_pmax25 Maximal PEP carboxylation rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for ac, aj, and a_p

Examples

c4 = C4VJPModel{Float64}();
c4 = C4VJPModel{Float64}(v_cmax25 = 30, v_pmax25 = 40, r_d25 = 1, colimit = true);

ClimaCache.AbstractPhotosynthesisMode — Type

abstract type AbstractPhotosynthesisMode

Hierarchy of AbstractSoilVC:

GCO₂Mode

PCO₂Mode

ClimaCache.GCO₂Mode — Type

struct GCO₂Mode <: ClimaCache.AbstractPhotosynthesisMode

An empty structure to signal the function to calculate photosynthetic rates based on leaf diffusive conductance to CO₂.

Examples

mode = GCO₂Mode();

ClimaCache.PCO₂Mode — Type

struct PCO₂Mode <: ClimaCache.AbstractPhotosynthesisMode

An empty structure to signal the function to calculate photosynthetic rates based on CO₂ partial pressure.

Examples

mode = PCO₂Mode();

ClimaCache.AbstractColimit — Type

abstract type AbstractColimit{FT<:AbstractFloat}

Hierarchy of AbstractColimit

MinimumColimit

QuadraticColimit

SerialColimit

ClimaCache.MinimumColimit — Type

struct MinimumColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Empty structure to indicate minimum colimitation.

Examples

col = MinimumColimit{Float64}();

ClimaCache.QuadraticColimit — Type

mutable struct QuadraticColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Structure to indicate quadratic colimitation.

Fields

CURVATURE::AbstractFloat

Curvature factor

Examples

col = QuadraticColimit{Float64}(0.98);

ClimaCache.SerialColimit — Type

struct SerialColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Empty structure to indicate serial colimitation.

Examples

col = SerialColimit{Float64}();

ClimaCache.AbstractTemperatureDependency — Type

abstract type AbstractTemperatureDependency{FT<:AbstractFloat}

Hierarchy of AbstractTemperatureDependency:

Arrhenius

ArrheniusPeak

Q10

ClimaCache.Arrhenius — Type

mutable struct Arrhenius{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An Arrhenius type struct using

\[Y_1 = Y_0 \cdot \exp \left( \dfrac{H_a}{R T_0} - \dfrac{H_a}{R T_1} \right)\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

ΔHA::AbstractFloat

Activation energy

ClimaCache.ArrheniusPeak — Type

mutable struct ArrheniusPeak{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An ArrheniusPeak type struct using

\[Y_1 = Y_0 \cdot \exp \left( \dfrac{H_a}{R T_0} - \dfrac{H_a}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

ΔHA::AbstractFloat

Activation energy

ΔHD::AbstractFloat

Deactivation energy

ΔSV::AbstractFloat

Entropy factor

ClimaCache.Q10 — Type

mutable struct Q10{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An Q10 type struct using

\[Y_1 = Y_0 \cdot Q_{10} ^ \dfrac{T_1 - T_0}{10}\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

Q_10::AbstractFloat

Power of Q10 correction

Radiation

ClimaCache.WaveLengthSet — Type

mutable struct WaveLengthSet{FT<:AbstractFloat}

Structure that stores wave length information.

Fields

IΛ_NIR::Vector{Int64}

Indicies of Λ_NIR in Λ

IΛ_PAR::Vector{Int64}

Indicies of Λ_PAR in Λ

IΛ_SIF::Vector{Int64}

Indicies of Λ_SIF in Λ

IΛ_SIFE::Vector{Int64}

Indicies of Λ_SIFE in Λ

NΛ::Int64

Number of wavelength bins

NΛ_PAR::Int64

Number of wavelength bins for PAR

NΛ_SIF::Int64

Number of wavelength bins for SIF

NΛ_SIFE::Int64

Number of wavelength bins for SIF excitation

WL_NIR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for NIR [nm]

WL_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for PAR [nm]

WL_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF emission [nm]

WL_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF excitation [nm]

SΛ::Vector{FT} where FT<:AbstractFloat

Standard wavelength (boundaries) [nm]

ΔΛ::Vector{FT} where FT<:AbstractFloat

Differential wavelength [nm]

ΔΛ_PAR::Vector{FT} where FT<:AbstractFloat

Differential wavelength for PAR [nm]

ΔΛ_SIFE::Vector{FT} where FT<:AbstractFloat

Differential wavelength for SIF excitation [nm]

Λ::Vector{FT} where FT<:AbstractFloat

Wavelength (bins) [nm]

Λ_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength bins for PAR [nm]

Λ_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF [nm]

Λ_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF excitation [nm]

ClimaCache.WaveLengthSet — Method

WaveLengthSet{FT}(swl::Vector = WAVELENGTHS; opti::String = OPTI_2021) where {FT<:AbstractFloat}

Constructor for WaveLengthSet, given

swl Standard wave length boundaries [nm], default is ClimaCache.WAVELENGTHS

opti Optical properties file path, default is ClimaCache.OPTI_2021

Examples

wls = WaveLengthSet{Float64}();
wls = WaveLengthSet{Float64}(collect(400:5:2500));
wls = WaveLengthSet{Float64}(collect(400:5:2500); opti=ClimaCache.OPTI_2017);

ClimaCache.HyperspectralRadiation — Type

mutable struct HyperspectralRadiation{FT<:AbstractFloat}

Structure that stores hyperspectral radiation information

Fields

e_direct::Vector{FT} where FT<:AbstractFloat

Direct radiation [mW m⁻² nm⁻¹]

e_diffuse::Vector{FT} where FT<:AbstractFloat

Diffuse radiation [mW m⁻² nm⁻¹]

ClimaCache.HyperspectralRadiation — Method

HyperspectralRadiation{FT}(wls::WaveLengthSet = WaveLengthSet{FT}(); file::String = FILE_SUN) where {FT<:AbstractFloat}

Constructor for HyperspectralRadiation, given

wls WaveLengthSet type struct that defines wavelength settings

file File path to solar radiation setting, default is ClimaCache.FILE_SUN

Examples

rad = HyperspectralRadiation{FT}();
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)));
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)); file = "");
rad = HyperspectralRadiation{FT}(WaveLengthSet{FT}(collect(400:50:2400)); file = ClimaCache.FILE_SUN);

ClimaCache.HyperspectralAbsorption — Type

mutable struct HyperspectralAbsorption{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflection and transmittance.

Fields

K_ANT::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of anthocynanin [-]

K_BROWN::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of senescent material (brown pigments) [-]

K_CAB::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of chlorophyll a and b [-]

K_CAR_V::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of violaxanthin carotenoid [-]

K_CAR_Z::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of zeaxanthin carotenoid [-]

K_CBC::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of carbon-based constituents [-]

K_H₂O::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of water [-]

K_LMA::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of dry matter [-]

K_PRO::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of protein [-]

K_PS::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of PS I and II [-]

NR::Vector{FT} where FT<:AbstractFloat

Refractive index [-]

ClimaCache.HyperspectralAbsorption — Method

HyperspectralAbsorption{FT}(wls::WaveLengthSet = WaveLengthSet{FT}(); opti::String = OPTI_2021) where {FT<:AbstractFloat}

Constructor for HyperspectralAbsorption, given

wls WaveLengthSet type structure

opti Path to leaf optical properties

Examples

ha = HyperspectralAbsorption{FT}();
ha = HyperspectralAbsorption{FT}(WaveLengthSet{FT}(collect(400:50:2400)));
ha = HyperspectralAbsorption{FT}(WaveLengthSet{FT}(collect(400:50:2400)); opti=ClimaCache.OPTI_2017);

Soil

ClimaCache.AbstractSoilVC — Type

abstract type AbstractSoilVC{FT<:AbstractFloat}

Hierarchy of AbstractSoilVC:

BrooksCorey

VanGenuchten

ClimaCache.BrooksCorey — Type

struct BrooksCorey{FT<:AbstractFloat} <: ClimaCache.AbstractSoilVC{FT<:AbstractFloat}

Brooks Corey soil parameters

Fields

TYPE::String

Soil type

B::AbstractFloat

Soil b

Ψ_SAT::AbstractFloat

Potential at saturation [MPa]

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

ClimaCache.VanGenuchten — Type

struct VanGenuchten{FT<:AbstractFloat} <: ClimaCache.AbstractSoilVC{FT<:AbstractFloat}

van Genuchten soil parameters

Fields

TYPE::String

Soil type

Α::AbstractFloat

Soil α is related to the inverse of the air entry suction, α > 0

N::AbstractFloat

Soil n is Measure of the pore-size distribution

M::AbstractFloat

Soil m = 1 - 1/n

Θ_SAT::AbstractFloat

Saturated soil volumetric water content

Θ_RES::AbstractFloat

Residual soil volumetric water content

ClimaCache.VanGenuchten — Method

VanGenuchten{FT}(name::String, α::Number, n::Number, θ_sat::Number, θ_res::Number) where {FT<:AbstractFloat}

Constructor for VanGenuchten, given

name Soil type name

α van Genuchten α

n van Genuchten n

θ_sat Saturated soil volumetric water content

θ_res Residual soil volumetric water content

Examples

vg = VanGenuchten{FT}("Test", FT(100), FT(2), FT(0.5), FT(0.1));

ClimaCache.VanGenuchten — Method

VanGenuchten{FT}(name::String) where {FT<:AbstractFloat}

Constructor for VanGenuchten, given

name Soil type name, need to be Sand, Loamy Sand, Sandy Loam, Loam, Sandy Clay Loam, Silt Loam, Silt (default), Clay Loam, Silty Clay Loam, Sandy Clay, Silty Clay, and Clay.

Examples

vg = VanGenuchten{FT}("");
vg = VanGenuchten{FT}("Loam");

SPAC

ClimaCache.Root — Type

mutable struct Root{FT<:AbstractFloat}

Structure to save root parameters

Fields

HS::ClimaCache.RootHydraulics{FT} where FT<:AbstractFloat

RootHydraulics type root hydraulic system

t::AbstractFloat

Current temperature

ClimaCache.Root — Method

Root{FT}(; ssm::Bool = true) where {FT<:AbstractFloat}

Construct a Root structure, given

ssm Whether the flow rate is at steady state

ClimaCache.Stem — Type

mutable struct Stem{FT<:AbstractFloat}

Structure to save stem parameters

Fields

HS::ClimaCache.StemHydraulics{FT} where FT<:AbstractFloat

StemHydraulics type stem hydraulic system

t::AbstractFloat

Current temperature

ClimaCache.Stem — Method

Stem{FT}(; ssm::Bool = true) where {FT<:AbstractFloat}

Construct a Stem structure, given

ssm Whether the flow rate is at steady state

ClimaCache.AbstractSPACSystem — Type

abstract type AbstractSPACSystem{FT<:AbstractFloat}

Hierarchy of AbstractSPACSystem:

MonoElementSPAC

MonoGrassSPAC

MonoPalmSPAC

MonoTreeSPAC

ClimaCache.MonoElementSPAC — Type

mutable struct MonoElementSPAC{FT} <: ClimaCache.AbstractSPACSystem{FT}

Struct for simplest SPAC system

Fields

LEAF::ClimaCache.Leaf{FT} where FT

Leaf system

ROOT::ClimaCache.Root{FT} where FT

Root system

STEM::ClimaCache.Stem{FT} where FT

Stem system

_krs::Vector

Relative hydraulic conductance

ClimaCache.MonoElementSPAC — Method

MonoElementSPAC{FT}(psm::String; ssm::Bool = true) where {FT<:AbstractFloat}

Construct a MonoElementSPAC type toy SPAC system, given

psm Photosynthesis model, must be C3, C4, or C3Cytochrome

ssm Whether the flow rate is at steady state

ClimaCache.MonoGrassSPAC — Type

mutable struct MonoGrassSPAC{FT} <: ClimaCache.AbstractSPACSystem{FT}

Struct for monospecies grass SPAC system

Fields

LEAVES::Array{ClimaCache.Leaf{FT}, 1} where FT

Leaf hydrualic system

LEAVES_INDEX::Vector{Int64}

Corresponding air layer per canopy layer

N_CANOPY::Int64

Number of canopy layers

N_ROOT::Int64

Number of root layers

ROOTS::Array{ClimaCache.Root{FT}, 1} where FT

Root hydraulic system

ROOTS_INDEX::Vector{Int64}

Corresponding soil layer per root layer

_fs::Vector

Flow rate per root layer

_ks::Vector

Conductances for each root layer at given flow

_ps::Vector

Pressure for each root layer at given flow

ClimaCache.MonoGrassSPAC — Method

MonoGrassSPAC{FT}(psm::String; zr::Number = -0.2, zc::Number = 0.5, zss::Vector = collect(0:-0.1:-1), zas::Vector = collect(0:0.05:1), ssm::Bool = true) where {FT<:AbstractFloat}

Construct a SPAC system for monospecies grass system, given

psm Photosynthesis model, must be C3, C4, or C3Cytochrome

zr Maximal root depth (negative value)

zc Maximal canopy height (positive value)

zss Vector of soil layer boundaries starting from 0

zas Vector of air layer boundaries starting from 0

ssm Whether the flow rate is at steady state

Examples

spac = MonoGrassSPAC{Float64}();
spac = MonoGrassSPAC{Float64}(zr = -0.3, zc = 1, zss = collect(0:-0.1:-1), zas = collect(0:0.05:1.01));

ClimaCache.MonoPalmSPAC — Type

mutable struct MonoPalmSPAC{FT} <: ClimaCache.AbstractSPACSystem{FT}

Struct for monospecies palm SPAC system (with trunk)

Fields

LEAVES::Array{ClimaCache.Leaf{FT}, 1} where FT

Leaf hydrualic system

LEAVES_INDEX::Vector{Int64}

Corresponding air layer per canopy layer

N_CANOPY::Int64

Number of canopy layers

N_ROOT::Int64

Number of root layers

ROOTS::Array{ClimaCache.Root{FT}, 1} where FT

Root hydraulic system

ROOTS_INDEX::Vector{Int64}

Corresponding soil layer per root layer

TRUNK::ClimaCache.Stem{FT} where FT

Trunk hydraulic system

_fs::Vector

Flow rate per root layer

_ks::Vector

Conductances for each root layer at given flow

_ps::Vector

Pressure for each root layer at given flow

ClimaCache.MonoPalmSPAC — Method

MonoPalmSPAC{FT}(psm::String; zr::Number = -1, zt::Number = 10, zc::Number = 12, zss::Vector = collect(0:-0.25:-2), zas::Vector = collect(0:0.2:13), ssm::Bool = true) where {FT<:AbstractFloat}

Construct a SPAC system for monospecies palm system, given

psm Photosynthesis model, must be C3 or C3Cytochrome

zr Maximal root depth (negative value)

zc Maximal canopy height (positive value)

zss Vector of soil layer boundaries starting from 0

zas Vector of air layer boundaries starting from 0

ssm Whether the flow rate is at steady state

Examples

spac = MonoPalmSPAC{Float64}();
spac = MonoPalmSPAC{Float64}(zr = -1, zt = 11, zc = 1, zss = collect(0:-0.1:-2), zas = collect(0:0.2:13));

ClimaCache.MonoTreeSPAC — Type

mutable struct MonoTreeSPAC{FT} <: ClimaCache.AbstractSPACSystem{FT}

Struct for monospecies tree SPAC system (with trunk and branches)

Fields

BRANCHES::Array{ClimaCache.Stem{FT}, 1} where FT

Branch hydraulic system

LEAVES::Array{ClimaCache.Leaf{FT}, 1} where FT

Leaf hydrualic system

LEAVES_INDEX::Vector{Int64}

Corresponding air layer per canopy layer

N_CANOPY::Int64

Number of canopy layers

N_ROOT::Int64

Number of root layers

ROOTS::Array{ClimaCache.Root{FT}, 1} where FT

Root hydraulic system

ROOTS_INDEX::Vector{Int64}

Corresponding soil layer per root layer

TRUNK::ClimaCache.Stem{FT} where FT

Trunk hydraulic system

_fs::Vector

Flow rate per root layer

_ks::Vector

Conductances for each root layer at given flow

_ps::Vector

Pressure for each root layer at given flow

ClimaCache.MonoTreeSPAC — Method

MonoTreeSPAC{FT}(psm::String; zr::Number = -1, zt::Number = 10, zc::Number = 12, zss::Vector = collect(0:-0.25:-2), zas::Vector = collect(0:0.2:13), ssm::Bool = true) where {FT<:AbstractFloat}

Construct a SPAC system for monospecies tree system, given

psm Photosynthesis model, must be C3, C4, or C3Cytochrome (note: there are C4 shrubs)

zr Maximal root depth (negative value)

zc Maximal canopy height (positive value)

zss Vector of soil layer boundaries starting from 0

zas Vector of air layer boundaries starting from 0

ssm Whether the flow rate is at steady state

Examples

spac = MonoTreeSPAC{Float64}();
spac = MonoTreeSPAC{Float64}(zr = -1, zt = 11, zc = 1, zss = collect(0:-0.1:-2), zas = collect(0:0.2:13));