ClimaCache

Air

ClimaCache.AirLayer — Type

mutable struct AirLayer{FT<:AbstractFloat}

Structure that stores air layer information

Fields

P_AIR::AbstractFloat

Atmospheric pressure [Pa]

P_O₂::AbstractFloat

O₂ partial pressure [Pa]

p_CO₂::AbstractFloat

CO₂ partial pressure [Pa]

ClimaCache.AirLayer — Method

AirLayer{FT}() where {FT<:AbstractFloat}

Constructor for AirLayer

Examples

air = AirLayer{Float64}();

Plant

ClimaCache.Leaf — Type

mutable struct Leaf{FT<:AbstractFloat}

Structure to save leaf parameters

Fields

BIO::ClimaCache.LeafBiophysics

LeafBiophysics type leaf biophysical parameters

PRC::ClimaCache.AbstractReactionCenter

AbstractReactionCenter type photosynthesis reaction center

PSM::ClimaCache.AbstractPhotosynthesisModel

AbstractPhotosynthesisModel type photosynthesis model

apar::AbstractFloat

Absorbed photosynthetically active radiation [μmol m⁻² s⁻¹]

g_H₂O_s::AbstractFloat

Stomatal conductance to water vapor [mol m⁻² s⁻¹]

t::AbstractFloat

Current leaf temperature

g_CO₂::AbstractFloat

Total leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

g_CO₂_b::AbstractFloat

Boundary leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

p_CO₂_i::AbstractFloat

Leaf internal CO₂ partial pressure [Pa]

p_CO₂_s::AbstractFloat

Leaf surface CO₂ partial pressure [Pa]

p_H₂O_sat::AbstractFloat

Saturation H₂O vapor pressure, need to update with temperature and leaf water pressure [Pa]

_t::AbstractFloat

Last leaf temperature. If different from t, then make temperature correction

ClimaCache.Leaf — Method

Leaf{FT}(psm::String, wls::WaveLengthSet{FT} = WaveLengthSet{FT}(); colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for Leaf, given

psm Photosynthesis model type, must be C3, C3Cytochrome, or C4

wls WaveLengthSet type structure that determines the dimensions of leaf parameters

Examples

leaf_c3 = Leaf{Float64}("C3");
leaf_c4 = Leaf{Float64}("C4");
leaf_cy = Leaf{Float64}("C3Cytochrome");
leaf_c3 = Leaf{Float64}("C3"; colimit = true);
leaf_c4 = Leaf{Float64}("C4"; colimit = true);
leaf_cy = Leaf{Float64}("C3Cytochrome"; colimit = true);
wls = WaveLengthSet{FT}(collect(400:10:2500));
leaf_c3 = Leaf{Float64}("C3", wls);
leaf_c4 = Leaf{Float64}("C4", wls);
leaf_cy = Leaf{Float64}("C3Cytochrome", wls);

ClimaCache.LeafBiophysics — Type

mutable struct LeafBiophysics{FT<:AbstractFloat}

Struct that contains leaf biophysical traits used to run leaf reflectance and transmittance.

Fields

MESOPHYLL_N::AbstractFloat

Leaf mesophyll structural parameter that describes mesophyll reflectance and transmittance

NDUB::Int64

Doubling adding layers

ant::AbstractFloat

Anthocyanin content [μg cm⁻²]

brown::AbstractFloat

Senescent material (brown pigments) fraction [-]

cab::AbstractFloat

Chlorophyll a and b content [μg cm⁻²]

car::AbstractFloat

Carotenoid content [μg cm⁻²]

cbc::AbstractFloat

Carbon-based constituents in lma [g cm⁻²]

f_zeax::AbstractFloat

Zeaxanthin fraction in Carotenoid (1=all Zeaxanthin, 0=all Violaxanthin) [-]

l_H₂O::AbstractFloat

Equivalent water thickness [cm]

lma::AbstractFloat

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

pro::AbstractFloat

Protein content in lma (pro = lma - cbc) [g cm⁻²]

k_all::Vector{FT} where FT<:AbstractFloat

Specific absorption coefficients of all materials

mat_b::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix backwards [-]

mat_f::Matrix{FT} where FT<:AbstractFloat

Fluorescence excitation matrix forwards [-]

α_cab::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll [-]

α_cabcar::Vector{FT} where FT<:AbstractFloat

Relative absorption by Chlorophyll+Carotenoid [-]

α_SW::Vector{FT} where FT<:AbstractFloat

Shortwave absorption, 1 .- ρSW .- τSW [-]

ρ_LW::AbstractFloat

Broadband thermal reflectance, related to blackbody emittance [-]

ρ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf reflectance [-]

τ_LW::AbstractFloat

Broadband thermal transmission, related to blackbody emittance [-]

τ_SW::Vector{FT} where FT<:AbstractFloat

Shortwave leaf transmission [-]

ClimaCache.LeafBiophysics — Method

LeafBiophysics{FT}(wls::WaveLengthSet{FT} = WaveLengthSet{FT}()) where {FT<:AbstractFloat}

Constructor for LeafBiophysics, given

wls WaveLengthSet type structure

Examples

lbio = LeafBiophysics{Float64}();
lbio = LeafBiophysics{Float64}(WaveLengthSet{Float64}(collect(400:50:2400)));

ClimaCache.VanDerTolFluorescenceModel — Type

struct VanDerTolFluorescenceModel{FT<:AbstractFloat}

Structure that stores van der Tol et al. (2014) fluorescence model parameters.

Fields

K_0::AbstractFloat

Fitting parameter K_0

K_A::AbstractFloat

Fitting parameter α

K_B::AbstractFloat

Fitting parameter β

ClimaCache.VanDerTolFluorescenceModel — Method

VanDerTolFluorescenceModel{FT}(drought::Bool = true) where {FT<:AbstractFloat}

Constructor for VanDerTolFluorescenceModel fluorescence model, given

drought If true, use parameters trained from drought stressed plant. Default is true.

Examples

vdt = VanDerTolFluorescenceModel{Float64}();
vdt = VanDerTolFluorescenceModel{Float64}(false);

ClimaCache.AbstractReactionCenter — Type

abstract type AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Hierachy of the AbstractReactionCenter

VJPReactionCenter

CytochromeReactionCenter

ClimaCache.VJPReactionCenter — Type

mutable struct VJPReactionCenter{FT<:AbstractFloat} <: ClimaCache.AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Fields

F_PSII::AbstractFloat

Fraction of absorbed light used by PSII ETR

FLM::ClimaCache.VanDerTolFluorescenceModel

Fluorescence model

K_D::AbstractFloat

Rate constant for thermal dissipation

K_F::AbstractFloat

Rate constant for fluorescence

K_P_MAX::AbstractFloat

Maximal rate constant for photochemistry

Φ_PSII_MAX::AbstractFloat

max PSII yield (knpqr=0, all RC open)

k_npq_rev::AbstractFloat

Reversible NPQ rate constant (initially zero)

k_npq_sus::AbstractFloat

Sustained NPQ rate constant (for seasonal changes, default is zero)

k_p::AbstractFloat

Rate constant for photochemistry

npq::AbstractFloat

Non-Photochemical quenching

ϕ_f::AbstractFloat

Fluorescence yield

ϕ_p::AbstractFloat

Photochemical yield

f_m::AbstractFloat

Dark adapted yield (Kp=0)

f_m′::AbstractFloat

Light adapted yield (Kp=0)

f_o::AbstractFloat

Dark-adapted fluorescence yield (Kp=max)

f_o′::AbstractFloat

Light-adapted fluorescence yield in the dark (Kp=max)

q_e::AbstractFloat

Energy quenching

q_p::AbstractFloat

Photochemical quenching

ClimaCache.VJPReactionCenter — Method

VJPReactionCenter{FT}() where {FT<:AbstractFloat}

Constructor of VJPReactionCenter

Examples

rc = VJPReactionCenter{Float64}();

ClimaCache.CytochromeReactionCenter — Type

mutable struct CytochromeReactionCenter{FT<:AbstractFloat} <: ClimaCache.AbstractReactionCenter{FT<:AbstractFloat}

Structure that stores reaction center information

Fields

F_PSI::AbstractFloat

Fraction of absorbed light used by PSI ETR

K_D::AbstractFloat

Rate constant of consititutive heat loss from the antennae [ns⁻¹]

K_F::AbstractFloat

Rate constant of fluorescence [ns⁻¹]

K_PSI::AbstractFloat

Rate constant of photochemistry for PS I [ns⁻¹]

K_PSII::AbstractFloat

Rate constant of photochemistry for PS II [ns⁻¹]

K_U::AbstractFloat

Rate constant of excitation sharing for PS II [ns⁻¹]

K_X::AbstractFloat

Rate constant of regulated heat loss via oxidized PS I center [s⁻¹]

Φ_PSI_MAX::AbstractFloat

Maximal PS I photochemical yield

ϵ_1::AbstractFloat

Weight factor that PSI fluorescence reaches sensor (after reabsorption)

ϵ_2::AbstractFloat

Weight factor that PSII fluorescence reaches sensor (after reabsorption)

ϕ_f::AbstractFloat

Fluorescence yield

ϕ_p::AbstractFloat

Photochemical yield

ClimaCache.CytochromeReactionCenter — Method

CytochromeReactionCenter{FT}() where {FT<:AbstractFloat}

Constructor of CytochromeReactionCenter

Examples

rc = CytochromeReactionCenter{Float64}();

ClimaCache.AbstractPhotosynthesisModel — Type

abstract type AbstractPhotosynthesisModel{FT<:AbstractFloat}

Hierachy of AbstractPhotosynthesisModel:

C3CytochromeModel

C3VJPModel

C4VJPModel

ClimaCache.C3CytochromeModel — Type

mutable struct C3CytochromeModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C3 Cytochrome photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

COLIMIT_J::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for J

EFF_1::AbstractFloat

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

EFF_2::AbstractFloat

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

TD_KC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kc temperature dependency

TD_KO::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Ko temperature dependency

TD_KQ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kq temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_Γ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Γ* temperature dependency

TD_ΗC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Η_C temperature dependency

TD_ΗL::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Η_L temperature dependency

b₆f::AbstractFloat

Total concentration of Cytochrome b₆f [μmol m⁻²]

k_q::AbstractFloat

Maximal turnover rate of Cytochrome b₆f [e⁻ s⁻¹]

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol e⁻ m⁻² s⁻¹]

j_psi::AbstractFloat

PSI electron transport rate after colimitation

k_c::AbstractFloat

RubisCO coefficient Kc [Pa]

k_m::AbstractFloat

Michaelis-Menten's coefficient [Pa]

k_o::AbstractFloat

RubisCO coefficient Ko [Pa]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

v_qmax::AbstractFloat

Maximal Cytochrome b₆f activity [μmol e⁻ m⁻² s⁻¹]

η::AbstractFloat

ratio between JP700 and JP680

η_c::AbstractFloat

Coupling efficiency of cyclic electron flow [mol ATP mol⁻¹ e⁻]

η_l::AbstractFloat

Coupling efficiency of linear electron flow [mol ATP mol⁻¹ e⁻]

γ_star::AbstractFloat

CO₂ compensation point with the absence of Rd [Pa]

ClimaCache.C3CytochromeModel — Method

C3CytochromeModel{FT}(; v_cmax25::Number = 50, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C3CytochromeModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for ac, aj, and a_p

Examples

cy = C3CytochromeModel{Float64}();
cy = C3CytochromeModel{Float64}(v_cmax25 = 30, r_d25 = 1, colimit = true);

ClimaCache.C3VJPModel — Type

mutable struct C3VJPModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C3 photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

COLIMIT_J::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for J

EFF_1::AbstractFloat

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

EFF_2::AbstractFloat

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

TD_JMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Jmax temperature dependency

TD_KC::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kc temperature dependency

TD_KO::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Ko temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_Γ::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Γ* temperature dependency

j_max25::AbstractFloat

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j::AbstractFloat

Electron transport [μmol m⁻² s⁻¹]

j_max::AbstractFloat

Maximal electron transport rate at leaf temperature [μmol m⁻² s⁻¹]

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

k_c::AbstractFloat

RubisCO coefficient Kc [Pa]

k_m::AbstractFloat

Michaelis-Menten's coefficient [Pa]

k_o::AbstractFloat

RubisCO coefficient Ko [Pa]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

γ_star::AbstractFloat

CO₂ compensation point with the absence of Rd [Pa]

ClimaCache.C3VJPModel — Method

C3VJPModel{FT}(; v_cmax25::Number = 50, j_max25::Number = 83.5, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C3VJPModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

j_max25 Maximal electron transport rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for j and ac, aj, and a_p

Examples

c3 = C3VJPModel{Float64}();
c3 = C3VJPModel{Float64}(v_cmax25 = 30, j_max25 = 50, r_d25 = 1, colimit = true);

ClimaCache.C4VJPModel — Type

mutable struct C4VJPModel{FT<:AbstractFloat} <: ClimaCache.AbstractPhotosynthesisModel{FT<:AbstractFloat}

Structure that stores C4 photosynthesis system information

Fields

COLIMIT_CJ::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ac and Aj => Ai

COLIMIT_IP::ClimaCache.AbstractColimit

AbstractColimit type colimitation method for Ai and Ap => Ag

TD_KPEP::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Kpep temperature dependency

TD_R::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type respiration temperature dependency

TD_VCMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vcmax temperature dependency

TD_VPMAX::ClimaCache.AbstractTemperatureDependency

AbstractTemperatureDependency type Vpmax temperature dependency

r_d25::AbstractFloat

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_cmax25::AbstractFloat

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

v_pmax25::AbstractFloat

Maximal PEP carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

a_c::AbstractFloat

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

a_gross::AbstractFloat

Gross photosynthetic rate [μmol m⁻² s⁻¹]

a_j::AbstractFloat

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

a_net::AbstractFloat

Net photosynthetic rate [μmol m⁻² s⁻¹]

a_p::AbstractFloat

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

e_to_c::AbstractFloat

Electron to CO₂ coefficient

j_pot::AbstractFloat

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

k_pep::AbstractFloat

PEP coefficient Kpep [Pa]

r_d::AbstractFloat

Respiration rate at leaf temperature [μmol m⁻² s⁻¹]

v_cmax::AbstractFloat

Maximal carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

v_pmax::AbstractFloat

Maximal PEP carboxylation rate at leaf temperature [μmol m⁻² s⁻¹]

ClimaCache.C4VJPModel — Method

C4VJPModel{FT}(; v_cmax25::Number = 50, v_pmax25::Number = 50, r_d25::Number = 0.75, colimit::Bool = false) where {FT<:AbstractFloat}

Constructor for C4VJPModel, given

v_cmax25 Maximal carboxylation rate at 298.15 K

v_pmax25 Maximal PEP carboxylation rate at 298.15 K

r_d25 Respiration rate at 298.15 K

colimit If true, use quadratic colimitations for ac, aj, and a_p

Examples

c4 = C4VJPModel{Float64}();
c4 = C4VJPModel{Float64}(v_cmax25 = 30, v_pmax25 = 40, r_d25 = 1, colimit = true);

ClimaCache.AbstractPhotosynthesisMode — Type

abstract type AbstractPhotosynthesisMode

Hierachy of AbstractSoilVC:

GCO₂Mode

PCO₂Mode

ClimaCache.GCO₂Mode — Type

struct GCO₂Mode <: ClimaCache.AbstractPhotosynthesisMode

An empty structure to signal the function to calculate photosynthetic rates based on leaf diffusive conductance to CO₂.

Examples

mode = GCO₂Mode();

ClimaCache.PCO₂Mode — Type

struct PCO₂Mode <: ClimaCache.AbstractPhotosynthesisMode

An empty structure to signal the function to calculate photosynthetic rates based on CO₂ partial pressure.

Examples

mode = PCO₂Mode();

ClimaCache.AbstractColimit — Type

abstract type AbstractColimit{FT<:AbstractFloat}

Hierachy of AbstractColimit

MinimumColimit

QuadraticColimit

SerialColimit

ClimaCache.MinimumColimit — Type

struct MinimumColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Empty structure to indicate minimum colimitation.

Examples

col = MinimumColimit{Float64}();

ClimaCache.QuadraticColimit — Type

mutable struct QuadraticColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Structure to indicate quadratic colimitation.

Fields

CURVATURE::AbstractFloat

Curvature factor

Examples

col = QuadraticColimit{Float64}(0.98);

ClimaCache.SerialColimit — Type

struct SerialColimit{FT<:AbstractFloat} <: ClimaCache.AbstractColimit{FT<:AbstractFloat}

Empty structure to indicate serial colimitation.

Examples

col = SerialColimit{Float64}();

ClimaCache.AbstractTemperatureDependency — Type

abstract type AbstractTemperatureDependency{FT<:AbstractFloat}

Hierachy of AbstractTemperatureDependency:

Arrhenius

ArrheniusPeak

Q10

ClimaCache.Arrhenius — Type

mutable struct Arrhenius{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An Arrhenius type struct using

\[Y_1 = Y_0 \cdot \exp \left( \dfrac{H_a}{R T_0} - \dfrac{H_a}{R T_1} \right)\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

ΔHA::AbstractFloat

Activation energy

ClimaCache.ArrheniusPeak — Type

mutable struct ArrheniusPeak{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An ArrheniusPeak type struct using

\[Y_1 = Y_0 \cdot \exp \left( \dfrac{H_a}{R T_0} - \dfrac{H_a}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

ΔHA::AbstractFloat

Activation energy

ΔHD::AbstractFloat

Deactivation energy

ΔSV::AbstractFloat

Entropy factor

ClimaCache.Q10 — Type

mutable struct Q10{FT<:AbstractFloat} <: ClimaCache.AbstractTemperatureDependency{FT<:AbstractFloat}

An Q10 type struct using

\[Y_1 = Y_0 \cdot Q_{10} ^ \dfrac{T_1 - T_0}{10}\]

Fields

T_REF::AbstractFloat

Reference temperature [K]

VAL_REF::AbstractFloat

Uncorrected vakye at reference temperature

Q_10::AbstractFloat

Power of Q10 correction

Radiation

ClimaCache.WaveLengthSet — Type

mutable struct WaveLengthSet{FT<:AbstractFloat}

Structure that stores wave length information.

Fields

IΛ_NIR::Vector{Int64}

Indicies of Λ_NIR in Λ

IΛ_PAR::Vector{Int64}

Indicies of Λ_PAR in Λ

IΛ_SIF::Vector{Int64}

Indicies of Λ_SIF in Λ

IΛ_SIFE::Vector{Int64}

Indicies of Λ_SIFE in Λ

NΛ::Int64

Number of wavelength bins

NΛ_PAR::Int64

Number of wavelength bins for PAR

NΛ_SIF::Int64

Number of wavelength bins for SIF

NΛ_SIFE::Int64

Number of wavelength bins for SIF excitation

WL_NIR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for NIR [nm]

WL_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength limits for PAR [nm]

WL_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF emission [nm]

WL_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength limits for SIF excitation [nm]

SΛ::Vector{FT} where FT<:AbstractFloat

Standard wavelength (boundaries) [nm]

ΔΛ::Vector{FT} where FT<:AbstractFloat

Differential wavelength [nm]

ΔΛ_PAR::Vector{FT} where FT<:AbstractFloat

Differential wavelength for PAR [nm]

ΔΛ_SIFE::Vector{FT} where FT<:AbstractFloat

Differential wavelength for SIF excitation [nm]

Λ::Vector{FT} where FT<:AbstractFloat

Wavelength (bins) [nm]

Λ_PAR::Vector{FT} where FT<:AbstractFloat

Wavelength bins for PAR [nm]

Λ_SIF::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF [nm]

Λ_SIFE::Vector{FT} where FT<:AbstractFloat

Wavelength bins for SIF excitation [nm]

ClimaCache.WaveLengthSet — Method

WaveLengthSet{FT}(swl::Vector=WAVELENGTHS; opti::String=OPTI_2021) where {FT<:AbstractFloat}

Constructor for WaveLengthSet, given

swl Standard wave length boundaries [nm], default is ClimaCache.WAVELENGTHS

opti Optical properties file path, default is ClimaCache.OPTI_2021

Examples

wls = WaveLengthSet{Float64}();
wls = WaveLengthSet{Float64}(collect(400:5:2500));
wls = WaveLengthSet{Float64}(collect(400:5:2500); opti=ClimaCache.OPTI_2017);